break runtime.go into smaller files
